+++ lib/crossbeams/layout/renderer/input.rb
@@ -107,7 +107,7 @@ module Crossbeams
         end
 
         def value
-          res = @page_config.form_object[@field_name]
+          res = form_object_value
           res = @page_config.form_values[@field_name] if @page_config.form_values
           if res.is_a?(BigDecimal) # TODO: read other frameworks to see best way of handling this...
             res.to_s('F')