+++ lib/crossbeams/layout/renderer/select.rb
@@ -22,7 +22,7 @@ module Crossbeams
           attrs << 'disabled="true"' if @field_config[:disabled] && @field_config[:disabled] == true
           attrs << 'required="true"' if @field_config[:required] && @field_config[:required] == true
           attrs << behaviours
-          sel = @field_config[:selected] || @page_config.form_object[@field_name]
+          sel = @field_config[:selected] || form_object_value
           sel = @page_config.form_values[@field_name] if @page_config.form_values
 
           disabled_option = find_disabled_option(sel, @field_config[:disabled_options])